requirements.txt
streamlit
hdbscan
sentence-transformers
transformers
umap-learn[parametric_umap]
seaborn
tf-keras==2.18.0
sklearn
datasets